1. add prop type for TooltipWithParser 2. rename filename  InputParametersTooltip to TooltipWithParser
