+++ packages/ui/src/ui-component/TooltipWithParser.js
@@ -1,6 +1,7 @@
 import { Info } from '@mui/icons-material';
 import { IconButton, Tooltip } from '@mui/material';
 import parser from 'html-react-parser';
+import PropTypes from 'prop-types';
 
 export const TooltipWithParser = ({ title }) => {
     return (
@@ -10,4 +11,8 @@ export const TooltipWithParser = ({ title }) => {
             </IconButton>
         </Tooltip>
     );
+};
+
+TooltipWithParser.propTypes = {
+    title: PropTypes.node
 };
